Refactor scheduler to use an idle watcher for sending signalChildJob()
The _changes reader now reads all avaliable data from the filehandle instead
of one line at a time.

The job-is-done processor maintians a list of child job IDs we need to
tell to decrement their waiting-on counters.  And idle watcher pulls items
off this queue one at a time and send the REST call to signalChildJob